generative_ai/AI_data_prep_assistant/requirements.txt
streamlit
pandas
numpy
scikit-learn
scipy
statsmodels
openai 
datasketch
datarobot
python-dotenv